Integrates new streamer Adds `test` command - allows to stream using new streamer and check success rate Changes configuration so for every service own VM gets created

+++ src/swarm.js
@@ -31,26 +31,44 @@ class Swarm {
     this._machines = null
   }
 
-  createMachines (opts, cb) {
-    const machinesCount = opts.machines.num || 3
-    const name = opts.name || 'testharness-' + shortid.generate()
-    const machine2type = this.mapMachinesTypes(opts)
-    console.log(machine2type)
+  createMachines (config, cb) {
+    this._updateMachines = config.updateMachines
+    this._installNodeExporter = config.installNodeExporter
+    this._installGoogleMonitoring = config.installGoogleMonitoring
+    const machinesCount = config.machines.num || 3
+    const name = config.name || 'testharness-' + shortid.generate()
+    const getMachineType = machineName => {
+      const cm = config.machines
+      if (cm.machine2serviceType.has(machineName)) {
+        switch (cm.machine2serviceType.get(machineName)) {
+        case 'streamer':
+          return cm.streamerMachineType
+        case 'broadcaster':
+          return cm.broadcasterMachineType
+        case 'orchestrator':
+          return cm.orchestratorMachineType
+        case 'transcoder':
+          return cm.transcoderMachineType
+        }
+      } else {
+        return cm.managerMachineType
+      }
+    }
 
-    if (Array.isArray(opts.machines.zones)) {
-      let numberOfZones = opts.machines.zones.length
+    if (Array.isArray(config.machines.zones)) {
+      let numberOfZones = config.machines.zones.length
       let machinesPerGroup = Math.floor(machinesCount / numberOfZones)
       let groups = {}
       for (let i = 0, j = 0; i < machinesCount - 1; i++) {
-        if (!groups[opts.machines.zones[j]]) {
-          groups[opts.machines.zones[j]] = []
+        if (!groups[config.machines.zones[j]]) {
+          groups[config.machines.zones[j]] = []
         }
 
-        groups[opts.machines.zones[j]].push({
+        groups[config.machines.zones[j]].push({
           name: `${name}-worker-${i + 1}`,
-          zone: opts.machines.zones[j],
-          machineType: machine2type.get(`${name}-worker-${i + 1}`) || opts.machines.machineType,
-          tags: opts.machines.tags || `${name}-cluster`
+          zone: config.machines.zones[j],
+          machineType: getMachineType(`${name}-worker-${i + 1}`),
+          tags: config.machines.tags || `${name}-cluster`
         })
 
         j = ++j % numberOfZones
@@ -60,9 +78,9 @@ class Swarm {
         (done) => {
           this.createMachine({
             name: `${name}-manager`,
-            zone: opts.machines.zones[0],
-            machineType: opts.machines.managerMachineType,
-            tags: opts.machines.tags || `${name}-cluster`
+            zone: config.machines.zones[0],
+            machineType: config.machines.managerMachineType,
+            tags: config.machines.tags || `${name}-cluster`
           }, done)
         },
         (done) => {
@@ -74,16 +92,16 @@ class Swarm {
         }
       ], (err) => {
         if (err) return cb(err)
-        this.setupGCEMonitoring(opts).then(() => cb(null), cb)
+        this.setupGCEMonitoring(config).then(() => cb(null), cb)
       })
     } else {
       parallel([
         (done) => {
           this.createMachine({
             name: `${name}-manager`,
-            zone: opts.machines.zone,
-            machineType: opts.machines.managerMachineType,
-            tags: opts.machines.tags || `${name}-cluster`
+            zone: config.machines.zone,
+            machineType: config.machines.managerMachineType,
+            tags: config.machines.tags || `${name}-cluster`
           }, done)
         },
         (done) => {
@@ -92,15 +110,15 @@ class Swarm {
             const  mName = `${name}-worker-${i + 1}`
             this.createMachine({
               name: mName,
-              zone: opts.machines.zone,
-              machineType: machine2type.get(mName) || opts.machines.machineType,
-              tags: opts.machines.tags || `${name}-cluster`
+              zone: config.machines.zone,
+              machineType: getMachineType(mName),
+              tags: config.machines.tags || `${name}-cluster`
             }, next)
           }, done)
         }
       ], (err) => {
         if (err) return cb(err)
-        this.setupGCEMonitoring(opts).then(() => cb(null), cb)
+        this.setupGCEMonitoring(config).then(() => cb(null), cb)
       })
     }
   }
@@ -114,7 +132,7 @@ class Swarm {
     const formattedName = gc.projectPath(PROJECT_ID);
     // console.log('Formatted name:', formattedName)
     const [groups] = await gc.listGroups({name: formattedName})
-    console.log('====== groups:', groups)
+    // console.log('====== groups:', groups)
     // const group = await gc.getGroup({name: formattedName})
     // console.log('====== group:', group)
     if (!groups.find(g => g.displayName === config.name)) {
@@ -126,7 +144,8 @@ class Swarm {
         },
       }
       const [resp] = await gc.createGroup(cgreq)
-      console.log(`Group ${config.name} created:`, resp)
+      // console.log(`Group ${config.name} created:`, resp)
+      console.log(`Group ${config.name} created:`)
     } else {
       console.log(`Group ${config.name} already exists.`)
     }
@@ -151,7 +170,7 @@ class Swarm {
       console.log(`ip of ${sn} is ${ip} cli port: ${po && po['7935']}, isGeth: ${isGeth}`)
       const checkName = `${config.name}-${sn} status`
       if (checks.find(c => c.displayName === checkName)) {
-        console.log(`Uptime check '${checkName} already exists.`)
+        // console.log(`Uptime check '${checkName} already exists.`)
         continue
       }
       const port = isGeth ? 8545 : +po['7935']
@@ -178,7 +197,7 @@ class Swarm {
           },
           contentMatchers,
       }})
-      console.log(`Uptime check for service ${sn} created:`, upResp)
+      // console.log(`Uptime check for service ${sn} created:`, upResp)
     }
     let nc = null
     if (config.email) {
@@ -267,7 +286,7 @@ class Swarm {
           name: formattedName,
           alertPolicy
         })
-        console.log('Created alert policy', JSON.stringify(ap, null, 2))
+        // console.log('Created alert policy', JSON.stringify(ap, null, 2))
       }
     }
     // await utils.remotelyExec(
@@ -282,7 +301,7 @@ class Swarm {
     const formattedName = gc.projectPath(PROJECT_ID);
     // console.log('Formatted name:', formattedName)
     const [groups] = await gc.listGroups({name: formattedName})
-    console.log('====== groups:', groups)
+    // console.log('====== groups:', groups)
     for (let group of groups) {
       if (group.displayName === config.name) {
         await gc.deleteGroup({name: group.name})
@@ -328,7 +347,7 @@ class Swarm {
     const formattedName = gc.projectPath(PROJECT_ID)
     // console.log('Formatted name:', formattedName)
     const [groups] = await gc.listGroups({name: formattedName})
-    console.log('====== groups:', groups)
+    // console.log('====== groups:', groups)
     for (let group of groups) {
       if (reserved.indexOf(group.displayName) === -1) {
         await gc.deleteGroup({name: group.name})
@@ -419,24 +438,30 @@ class Swarm {
   }
 
   async setupMachine(machine, zone) {
-    await utils.remotelyExec(machine, zone,
-      `sudo curl -sSO https://dl.google.com/cloudagents/install-monitoring-agent.sh && sudo bash install-monitoring-agent.sh`
-    )
-    await utils.remotelyExec(machine, zone, `sudo apt-get update && sudo apt-get upgrade -y`)
-    console.log(`=============== apt updated`)
-    await utils.remotelyExec(machine, zone,
-       `sudo apt-get install -y python3-pip && sudo apt autoremove -y && \
-       sudo pip3 install ansible && \
-       ansible-galaxy install cloudalchemy.node-exporter && \
-       cat <<-SHELL_SCREENRC > $HOME/nodepb.yml
-- hosts: 127.0.0.1
-  connection: local
-  become: yes
-  roles:
-    - cloudalchemy.node-exporter
-SHELL_SCREENRC`
-    )
-    await utils.remotelyExec(machine, zone, 'sudo ansible-playbook  nodepb.yml')
+    if (this._installGoogleMonitoring) {
+      await utils.remotelyExec(machine, zone,
+        `sudo curl -sSO https://dl.google.com/cloudagents/install-monitoring-agent.sh && sudo bash install-monitoring-agent.sh`
+      )
+    }
+    if (this._updateMachines) {
+      await utils.remotelyExec(machine, zone, `sudo apt-get update && sudo apt-get upgrade -y`)
+      console.log(`=============== apt updated`)
+    }
+    if (this._installNodeExporter) {
+      await utils.remotelyExec(machine, zone,
+        `sudo apt-get install -y python3-pip && sudo apt autoremove -y && \
+        sudo pip3 install ansible && \
+        ansible-galaxy install cloudalchemy.node-exporter && \
+        cat <<-SHELL_SCREENRC > $HOME/nodepb.yml
+  - hosts: 127.0.0.1
+    connection: local
+    become: yes
+    roles:
+      - cloudalchemy.node-exporter
+  SHELL_SCREENRC`
+      )
+      await utils.remotelyExec(machine, zone, 'sudo ansible-playbook  nodepb.yml')
+    }
   }
 
   async setupMachine2(machine, zone) {
@@ -480,6 +505,7 @@ SHELL_SCREENRC`
   }
 
   getPubIP (machineName, cb) {
+    // console.log(`getPubIP(${machineName})`)
     return new Promise((resolve, reject) => {
       exec(`docker-machine ip ${machineName}`, (err, ip) => {
         if (err) {
@@ -590,29 +616,6 @@ SHELL_SCREENRC`
     }
     await this.createSwarmCreateMachines(config)
     return false
-  }
-
-  mapMachinesTypes (config) {
-    const m = new Map()
-    const cm = config.machines
-    if (!cm.orchestartorsMachines || !cm.broadcastersMachines) {
-      for (let i = 0; i < cm.num-1; i++) {
-        m.set(`${config.name}-worker-${i+1}`, cm.machineType)
-      }
-    } else {
-      if (cm.orchestartorsMachines + cm.broadcastersMachines + 1 > cm.num) {
-        console.log(chalk.red('Error in config - total number of machines should be greater than number of orchestrator\'s and broadcaster\'s machines.'))
-        process.exit(11)
-      }
-      const bcs = cm.orchestartorsMachines
-      const sts = bcs + cm.broadcastersMachines
-      for (let i = 0; i < cm.num-1; i++) {
-        const mt = i >= sts ? cm.streamerMachineType :
-          i >= bcs ? cm.broadcasterMachineType : cm.machineType
-        m.set(`${config.name}-worker-${i+1}`, mt)
-      }
-    }
-    return m
   }
 
   createSwarmCreateMachines (config) {
@@ -1069,6 +1072,14 @@ SHELL_SCREENRC`
         })
       }
     })
+  }
+
+  static async getManagerIP(configName) {
+    const swarm = new Swarm(configName)
+    // const ri = await swarm.getRunningMachinesList(configName)
+    // console.log(`running machines: "${ri}"`)
+    const ip = await swarm.getPubIP(`${configName}-manager`)
+    return ip
   }
 
   static async getPublicIPOfService (parsedCompose, serviceName) {
